requirements.txt
Flask
flask-cors
web3
pyjwt
cachelib
python-dotenv